// src/models/enums/InterventionSector.ts
export enum InterventionSector {
    PROTECTION = "Protection",
    NUTRITION = "Nutrition",
    LOGISTICS = "Logistique",
    WASH = "Eau, Hygiène et Assainissement (WASH)",
    HEALTH = "Santé",
    FOOD_SECURITY = "Sécurité alimentaire",
    SHELTER = "Abris",
    EDUCATION = "Éducation",
    INTER_AGENCY_COORDINATION = "Coordination inter-agence",
    OTHER = "Autre"
}